add debug for rate limit metrics test
