+++ src/meta_protocol_proxy/filters/local_ratelimit/stats.h
@@ -1,0 +1,40 @@
+#pragma once
+
+#include <string>
+
+#include "envoy/stats/scope.h"
+#include "envoy/stats/stats_macros.h"
+
+namespace Envoy {
+namespace Extensions {
+namespace NetworkFilters {
+namespace MetaProtocolProxy {
+namespace LocalRateLimit {
+
+/**
+ * All local rate limit stats. @see stats_macros.h
+ */
+#define ALL_LOCAL_RATE_LIMIT_STATS(COUNTER)                                                        \
+  COUNTER(rate_limited)                                                                            \
+  COUNTER(ok)
+
+/**
+ * Struct definition for all local rate limit stats. @see stats_macros.h
+ */
+struct LocalRateLimitStats {
+  ALL_LOCAL_RATE_LIMIT_STATS(GENERATE_COUNTER_STRUCT)
+
+  static LocalRateLimitStats generateStats(const std::string& prefix, Stats::Scope& scope) {
+    const std::string final_prefix = prefix + ".local_rate_limit";
+    std::cout << final_prefix << std::endl << std::endl;
+    std::cout << &scope << std::endl << std::endl;
+    return {ALL_LOCAL_RATE_LIMIT_STATS(POOL_COUNTER_PREFIX(scope, final_prefix))};
+  }
+};
+
+} // namespace LocalRateLimit
+} // namespace MetaProtocolProxy
+} // namespace NetworkFilters
+} // namespace Extensions
+} // namespace Envoy
+
