+++ test/local-ratelimit/test.sh
@@ -7,8 +7,8 @@ docker run -d --network host --name client --env helloServer=localhost --env mod
 docker run -d -p 9091:9090 --name server aeraki/thrift-sample-server
 kill `ps -ef | awk '/bazel-bin\/envoy/{print $2}'`
 #$BASEDIR/../../bazel-bin/envoy -c $BASEDIR/test-full-match-condition.yaml &
-#$BASEDIR/../../bazel-bin/envoy -c $BASEDIR/test-only-global-token.yaml &
+$BASEDIR/../../bazel-bin/envoy -c $BASEDIR/test-only-global-token.yaml &
 #$BASEDIR/../../bazel-bin/envoy -c $BASEDIR/test-full-no-matched-condition.yaml &
 #$BASEDIR/../../bazel-bin/envoy -c $BASEDIR/test-no-global-bucket-match-condition.yaml &
-$BASEDIR/../../bazel-bin/envoy -c $BASEDIR/test-no-global-bucket-no-matched-condition.yaml &
+#$BASEDIR/../../bazel-bin/envoy -c $BASEDIR/test-no-global-bucket-no-matched-condition.yaml &
 docker logs -f client
